Fixes for build warnings

+++ src/sbs/globals.cpp
@@ -265,7 +265,7 @@ float Round(float number, int decimal_places)
 	if (decimal_places <= 0)
 		return floorf(number + 0.5f);
 
-	float multiplier = powf(10.0f, decimal_places);
+	float multiplier = powf(10.0f, (float)decimal_places);
 	float rounded = floorf((number * multiplier) + 0.5f) / multiplier;
 	return rounded;
 }
@@ -299,7 +299,7 @@ bool IsBoolean(std::string string)
 
 float ToFloat(const std::string &string)
 {
-	return atof(string.c_str());
+	return (float)atof(string.c_str());
 }
 
 int ToInt(const std::string &string)
